Add authorization middleware to the groups route
- also added the bearer token to the header within axios in the web client

+++ server/api/routes/groupsRouter.js
@@ -1,13 +1,14 @@
 const express = require('express');
-const responses = require('../util/responses');
 const router = express.Router();
 
 const groupsModel = require('../model/groupsModel');
+const { authenticateToken } = require('../util/tokenHelper');
+const responses = require('../util/responses');
 
 // Create a new group
-router.post('/', (req, res, next) => {
+router.post('/', authenticateToken, (req, res, next) => {
     const newGroup = req.body;
-    const groupOwnerId = 1; // temporary for now, will update this with auth once it is setup
+    const groupOwnerId = req.user.userID;
 
     if (!newGroup.groupName) {
         res.status(responses.NOT_FOUND);
@@ -34,8 +35,8 @@ router.post('/', (req, res, next) => {
 });
 
 // Get all groups
-router.get('/', (req, res, next) => {
-    const userId = 1; // temporary for now, will update this with auth once it is setup
+router.get('/', authenticateToken, (req, res, next) => {
+    const userId = req.user.userID;
     if (!userId) {
         res.status(responses.NOT_FOUND);
         res.send({ error: `userId is required.`});
@@ -65,14 +66,21 @@ router.get('/:groupId', (req, res, next) => {
 });
 
 // Get group information from groupId
-router.delete('/:groupId', (req, res, next) => {
+router.delete('/:groupId', authenticateToken, (req, res, next) => {
     const { groupId } = req.params;
+    const userID = req.user.userID;
     return groupsModel.getGroupFromGroupId(groupId)
         .then((result) => {
             if(result.length > 0) {
-                return groupsModel.deleteGroup(groupId).then(() => {
-                    res.status(responses.SUCCESS).json({deletedGroup: result[0]});
-                });
+                if(result[0].GroupOwnerId != userID) {
+                    res.status(responses.FORBIDDEN);
+                    res.send({ error: `Need to be group owner to delete group.`});
+                }
+                else {
+                    return groupsModel.deleteGroup(groupId).then(() => {
+                        res.status(responses.SUCCESS).json({deletedGroup: result[0]});
+                    });
+                }
             }
             else {
                 res.status(responses.NOT_FOUND);
